作业系统代码上传

+++ service/src/main/java/com/homework/mapper/HomeworkClassMapper.java
@@ -6,5 +6,5 @@ import com.homework.model.HomeworkClass;
  * Created by Administrator on 2017/4/7 0007.
  */
 public interface HomeworkClassMapper {
-    int insertHomeworkClass(HomeworkClass homeworkClass);
+    Long insertHomeworkClass(HomeworkClass homeworkClass);
 }
